// frontend/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        MonthDaySummaryResponse: {
            /**
             * Day
             * Format: date
             */
            day: string;
            /** Completed Count */
            completed_count: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}